+++ src/backend/marsha/core/views.py
@@ -6,7 +6,10 @@ import uuid
 
 from django.conf import settings
 from django.core.cache import cache
-from django.core.exceptions import ImproperlyConfigured
+from django.core.exceptions import (
+    ImproperlyConfigured,
+    ValidationError as DjangoValidationError,
+)
 from django.templatetags.static import static
 from django.utils.decorators import method_decorator
 from django.views.decorators.clickjacking import xframe_options_exempt
@@ -15,6 +18,8 @@ from django.views.generic import View
 from django.views.generic.base import TemplateResponseMixin, TemplateView
 
 from pylti.common import LTIException
+from rest_framework.exceptions import ValidationError as DRFValidationError
+from rest_framework.views import exception_handler as drf_exception_handler
 from rest_framework_simplejwt.tokens import AccessToken
 from waffle import switch_is_active
 
@@ -28,6 +33,26 @@ from .utils.react_locales_utils import react_locale
 
 
 logger = getLogger(__name__)
+
+
+def exception_handler(exc, context):
+    """Handle Django ValidationError as an accepted exception.
+
+    For the parameters, see ``exception_handler``
+    This code comes from twidi's gist:
+    https://gist.github.com/twidi/9d55486c36b6a51bdcb05ce3a763e79f
+    """
+    if isinstance(exc, DjangoValidationError):
+        if hasattr(exc, "message_dict"):
+            detail = exc.message_dict
+        elif hasattr(exc, "message"):
+            detail = exc.message
+        elif hasattr(exc, "messages"):
+            detail = exc.messages
+
+        exc = DRFValidationError(detail=detail)
+
+    return drf_exception_handler(exc, context)
 
 
 @method_decorator(csrf_exempt, name="dispatch")